third_party/WebKit/LayoutTests/platform/mac-mac10.10/external/wpt/IndexedDB/idbindex-rename-errors-expected.txt

This is a testharness.js-based test.
FAIL IndexedDB deleted index rename throws assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB index rename throws in a readonly transaction assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB index rename throws in a readwrite transaction assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB index rename throws in an inactive transaction assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB index rename to the name of another index throws assert_equals: Expected success event, but got error event instead expected "success" but got "error"
FAIL IndexedDB index rename handles exceptions when stringifying names assert_equals: Expected success event, but got error event instead expected "success" but got "error"
Harness: the test ran to completion.
